Links to official packages now point back to AL site.
Previously if the package was not in community or unsupported, the link didn't point anywhere. Now it uses http://archlinux.org/packages/search/$pkgname. If there's a better way, someone needs to let me know. As far as I know, there's no sure way to link directly to a package without knowing the pkgid.

Signed-off-by: tardo <tardo@nagi-fanboi.net>