Several functions added to web/lib/acctfuncs.inc Weeere back!
try_login() to login users

valid_username() checks if a new username fits criteria

valid_user() checks if the user exists in the database

good_passwd() only checks for minimum password length for now.
can be later expanded to tell a user to make a stronger password.

valid_passwd() checks if the password for the specified user is correct

user_suspended() checks if the user is suspended (or not)

user_delete() deletes a user (it doesn't orphan PKGs yet though)

user_is_privileged() returns privilege level User (0) TU (2) Dev (3) of
user ID. 0 is used for a regular user for ease in conditionals.

Also:	Enforce proper usernames on account creation or editing
	Fix bug where $SUPPORTED_LANGS needs to be reset on account creation
	Fix bug where an account could be created with an empty passwd
	Display (required) beside password fields on account creation
	Enforce good_passwd() on account creation

	TUs and Devs can edit a user to have a username that doesn't conform to
	the standard valid_username(). This is to allow them to edit old
	accounts without messing up the user name.

Signed-off-by: Loui Chang <louipc.ist@gmail.com>