Adds support for TUs and devs to disown packages they do not own
This is to add support for either devs or TUs to disown packages
whether they own them or not. I know of countless times where I or
another TU have been asked to orphan packages for someone and end up
having to adopt the package first and then disown it, this gets really
tedious for more than one package. As far as I can tell there's no
other way to disown packages you don't own and if there is at least
this is a more obvious way, pretty sure I didn't leave anything out in
the patch.

Signed-off-by: Callan Barrett <wizzomafizzo@gmail.com>