Bug in uploading PKGBUILDs and aur schema
This fixes a bug where TUs and devs couldn't upload packages because
the query would have an extra column in it to mark the pkgbuild safe
automatically, guessing it got missed when the safe flags were
removed. Also fixes a screw up I made with the schema file when I
added the tables for the voting app.

Signed-off-by: Callan Barrett <wizzomafizzo@gmail.com>