Allow pkgsubmit to accept any tar* file
Instead of restricting to gz and bz2, allow just plain tar and any other
format tar can understand to be uploaded. I make the mistake all the time of
forgetting to pass -z to tar when making something to upload, and there is
no real reason to exclude plain tar files.

Signed-off-by: Dan McGee <dan@archlinux.org>