Add a new DEFAULT_LANG constant.
DEFAULT_LANG will essentially be used to specify what language
strings are initially written in.

This will eliminate the need for English translation arrays in
AUR and make adding or changing the English strings a lot easier.

DEFAULT_LANG may be required for strings to display properly.

Also change the output when a translation isn't found.
Eliminate the <b> which can cause validation errors depending
on where the string is placed.

Signed-off-by: Loui Chang <louipc.ist@gmail.com>