Use non-persistant connections
There is really no need to use persistent connections to the database in
this day and age. Most PHP development guides recommend against it, and the
new mysqli interface doesn't even include the functionality.

Add a matching but currently unused db_disconnect() function while we are at
it. Reference counting will cover us for the most part, however.

Signed-off-by: Dan McGee <dan@archlinux.org>
Signed-off-by: Loui Chang <louipc.ist@gmail.com>