renamed rss2 -> rss
changed file name, as the 2 is relatively meaningless.
I think at one time it qualified it as a rss2.0 feed, but the
distinction is valueless at this point, since the generated xml
itself declares the rss version.

make sure to set an apache rewrite rule for any production systems
that require it.

Signed-off-by: Loui Chang <louipc.ist@gmail.com>